Changed folder structure, changed API and implemented OOP.
As said in #3.

Code is kind of bloated in tiles.py and OOP doesn't kinda work between panorama and tiles even if they're in the same folder, which is kinda weird.

I also forgot to document some functions in both of these scripts so there's that.

The old script is still there (in the folder old) and I'll still keep it until I implement Tweepy, since I forgot how it works again haha.